Properties returned by get_file_metadata() and get_images_metadata() are now dictionaries.
Any code that makes use of get_file_metadata() and get_images_metadata() should be reviewed (see xisfmeta.py example for reference). The dictionary returned by get_file_metadata(), and the 'XISFProperties' dictionary of the output of get_images_metadata() now contain a dictionary describing each property, i.e., it holds all the attributes of the property, instead of just the property value.
As the property attribute 'type' is preserved, _property_types is now unnecessary. This better handles any custom property not described in the XISF specification.
TODO: For now, the values are not converted to python native types (they are preserved as strings). Recall that only scalar, string and timepoint properties are supported; any other type, and the rest of the core elements (except FITSKeywords and Image, of course), are not yet planned.
Added get_metadata_xml() to provide access to the XML header.